bc4/requirements.txt
# Required packages for BC4 Procurement Planning Agent
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
fpdf2>=2.7.0
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
bcrypt>=4.0.0
python-multipart>=0.0.6
